Chores on the Blockchain
This project introduces the concept of tokenized rewards to incentivize completion of task on a Trello / Kanban style task board. Perfect for families and small teams.

Parents add task which their kids complete. Once parents approve the completed task. The kids or employees are paid. We use uPort for authentication and Colony for task creation and rating to release funds.

This task board can run on your web enabled family fridge. This includes Samsung refrigerators which have an LCD display with a webkit based web browser.

[Video Demo](https://www.youtube.com/watch?v=vg4r8eRJZmU)